# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: OpenLight Gateway API
  version: 1.0.0
  description: Proposed v1 contract pending independent review. Local-first normalized
    lighting API. See docs/API.md for semantic validation, ordering, idempotency,
    authentication and reconciliation.
servers:
- url: /api/v1
security:
- bearerAuth: []
paths:
...
  /rooms:
    get:
      operationId: get_rooms
      summary: List rooms
      responses:
        '200':
          description: List rooms
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
                required:
                - items
                additionalProperties: false
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    post:
      operationId: post_rooms
      summary: Create room
      responses:
        '201':
          description: Create room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreate'
...
  /groups:
    get:
      operationId: get_groups
      summary: List groups
      responses:
        '200':
          description: List groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                required:
                - items
                additionalProperties: false
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    post:
      operationId: post_groups
      summary: Create group
      responses:
        '201':
          description: Create group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
...
  /scenes:
    get:
      operationId: get_scenes
      summary: List scenes
      responses:
        '200':
          description: List scenes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scene'
                required:
                - items
                additionalProperties: false
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    post:
      operationId: post_scenes
      summary: Create scene
      responses:
        '201':
          description: Create scene
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneCreate'
...
  /presets:
    get:
      operationId: get_presets
      summary: List presets
      description: |
        Presets are named animated palettes that supported bulbs (currently Feit/Tuya) play by themselves from a
        single command. See docs/PRESETS.md for the full guide. The complete collection is returned, in creation order.
      responses:
        '200':
          description: List presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Preset'
                required:
                - items
                additionalProperties: false
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    post:
      operationId: post_presets
      summary: Create preset
      description: |
        Creates a preset and immediately advertises it as an `effects` effect ID (the preset `id`) on every device whose
        adapter supports native presets. No gateway restart is needed. Apply it with
        `POST /devices/{id}/commands` and `{"state":{"effect":"<preset id>"}}`.
      responses:
        '201':
          description: Preset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preset'
          headers:
            Location:
              schema:
                type: string
              description: Created resource URL
            ETag:
              schema:
                type: string
              description: Opaque quoted configuration revision
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresetCreate'
...
  /scenes/{id}/activate:
    post:
      operationId: post_scenes_id_activate
      summary: Activate scene
      responses:
        '202':
          description: Activate scene
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '404':
          $ref: '#/components/responses/404'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneActivate'
...
  /effects/{id}/start:
    post:
      operationId: post_effects_id_start
      summary: Start effect execution
      responses:
        '202':
          description: Start effect execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '404':
          $ref: '#/components/responses/404'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EffectStart'
...
  /effect-runs/{id}/stop:
    post:
      operationId: post_effect_runs_id_stop
      summary: Stop effect execution
      responses:
        '202':
          description: Stop effect execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '404':
          $ref: '#/components/responses/404'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Empty'
  /discovery/start:
    post:
      operationId: post_discovery_start
      summary: Start adapter discovery
      responses:
        '202':
          description: Start adapter discovery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
          headers:
            Location:
              schema:
                type: string
              description: Created resource or operation URL
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id001
        description: Persisted 24h minimum; same principal/method/path/body returns
          original response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryStart'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Room:
      type: object
      properties:
        id: *id001
        name: &id013
          type: string
          minLength: 1
      required:
      - id
      - name
      additionalProperties: false
      examples:
      - id: 1c2bd3ee-7334-400e-8bc1-96c02bf16b45
        name: Office
    Group:
      type: object
      properties:
        id: *id001
        name: &id015
          type: string
          minLength: 1
        deviceIds: &id016
          type: array
          items: *id001
          uniqueItems: true
      required:
      - id
      - name
      - deviceIds
      additionalProperties: false
      examples:
      - id: 1c2bd3ee-7334-400e-8bc1-96c02bf16b45
        name: Evening lights
        deviceIds:
        - 1c2bd3ee-7334-400e-8bc1-96c02bf16b45
...
    RoomCreate:
      type: object
      properties: &id014
        name: *id013
      required:
      - name
      additionalProperties: false
      description: Creation returns 422 resource_limit at the v1 installation limit
        of 10,000 resources per collection.
...
    GroupCreate:
      type: object
      properties: &id017
        name: *id015
        deviceIds: *id016
      required:
      - name
      - deviceIds
      additionalProperties: false
      description: Creation returns 422 resource_limit at the v1 installation limit
        of 10,000 resources per collection.
...
  responses:
    '400':
      description: Malformed syntax or missing/invalid Idempotency-Key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
    '401':
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
    '403':
      description: Forbidden or disallowed Origin
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
...
    '409':
      description: Conflict, offline device, active effect or key conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
...
    '413':
      description: Request body too large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
    '415':
      description: Unsupported media type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
    '422':
      description: Validation/capability failure or no eligible targets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
...
    '429':
      description: Rate/queue limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
      headers:
        Retry-After:
          schema:
            type: string
          description: Seconds until retry is permitted
    '500':
      description: Redacted internal failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027
    '503':
      description: Gateway or adapter unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: *id027